Added the data used for the hand_coded test with expected naming convention
We created a DE_0 folder in the test_data folder containing the data
from the two patients that were hand-coded. The file names are now what
the code is expecting for DE_1 through DE_20 of the real CMS data. This
enables quick testing, and comparison against the hand coded outputs.