ETL-CMS version 1.0
A complete overhaul of the project was performed to make the ETL
feature-complete for the OMOP CDMv5, and available for public use. All
tables were implemented to the extent that valid data was available
Extensive documentation was created for running the ETL, creating
database tables, and loading the data. Credit was attributed to known
contributors.